requirements.txt
amaranth-yosys==0.35.0.0.post81
git+https://github.com/amaranth-lang/amaranth@ab6503e352825b36bb29f1a8622b9e98aac9a6c6
black==23.3.0
pyright==1.1.332
pytest==7.2.2
amaranth_boards
